refactor: refactor out descriptor provider from 360 image provider and compose the functionality
